improved pin rotation

+++ fluorender/FluoRender/VRenderView.h
@@ -1002,11 +1002,11 @@ private:
 
 	//get mouse point in 3D
 	//mode: 0-maximum with original value; 1-maximum with transfered value; 2-accumulated with original value; 3-accumulated with transfered value
-	double GetPointVolume(Point &mp, int mx, int my, VolumeData* vd, int mode, bool use_transf, double thresh = 0.5);
-	double GetPointVolumeBox(Point &mp, int mx, int my, VolumeData* vd, bool calc_mats=true);
-	double GetPointVolumeBox2(Point &p1, Point &p2, int mx, int my, VolumeData* vd);
-	double GetPointPlane(Point &mp, int mx, int my, Point *planep=0, bool calc_mats=true);
-	Point* GetEditingRulerPoint(int mx, int my);
+	double GetPointVolume(Point &mp, double mx, double my, VolumeData* vd, int mode, bool use_transf, double thresh = 0.5);
+	double GetPointVolumeBox(Point &mp, double mx, double my, VolumeData* vd, bool calc_mats=true);
+	double GetPointVolumeBox2(Point &p1, Point &p2, double mx, double my, VolumeData* vd);
+	double GetPointPlane(Point &mp, double mx, double my, Point *planep=0, bool calc_mats=true);
+	Point* GetEditingRulerPoint(double mx, double my);
 
 	//brush sets
 	void ChangeBrushSetsIndex();
